+++ src/game_state/__init__.py
@@ -9,7 +9,7 @@ A utility package for pygame to manage multiple screens.
 
 """
 
-__version__ = "2.0.3"
+__version__ = "2.1.0b"
 __title__ = "game-state"
 __author__ = "Jiggly-Balls"
 __license__ = "MIT"